+++ index.d.ts
@@ -17,11 +17,18 @@ interface ReactTreeProps {
     data?: DataType,
 
     /**
-     * Triggered every time a node item is clicked. 
+     * Triggered every time a node item is clicked (left clicked). 
      * It receives as argurments the `id` and the `value` of the data for the node clicked.
      * The isLeafNode is true if, well, the node clicked is a leaf node.
      */
     onNodeClick?: (nodeId:string, nodeValue:any, isLeafNode:boolean) => any,
+
+    /**
+     * Triggered every time a node item is right-clicked. 
+     * It receives as argurments the `id` and the `value` of the data for the node clicked.
+     * The isLeafNode is true if, well, the node clicked is a leaf node.
+     */
+    onNodeRightClick?: (nodeId:string, nodeValue:any, isLeafNode:boolean) => any,
 
     /**
      * Provides a custom component to be used in rendering a parent node.